requirements.txt
discord.py~=2.4.0
yt-dlp~=2024.11.18
ffmpeg-python
PyNaCl
asyncio
matplotlib~=3.9.2
feedparser~=6.0.11
requests~=2.32.3
beautifulsoup4~=4.12.3
aiohttp~=3.11.6
selenium~=4.26.1
webdriver-manager==3.8.5
scikit-learn~=1.5.2
fuzzywuzzy~=0.18.0